+++ src/app/store/admin/admin.actions.ts
@@ -9,6 +9,7 @@ import {
   License,
   PublisherInfo,
   RejectedDomainsResponse,
+  UserBanDetails,
   UserInfo,
 } from 'models/settings.model'
 
@@ -70,6 +71,19 @@ export const GET_INDEX_FAILURE = 'Get index failure';
 export const GET_LICENSE = 'Get license';
 export const GET_LICENSE_SUCCESS = 'Get license success';
 export const GET_LICENSE_FAILURE = 'Get license failure';
+
+export const BAN_USER = 'Ban user';
+export const BAN_USER_SUCCESS = 'Ban user success';
+export const BAN_USER_FAILURE = 'Ban user failure';
+
+export const UNBAN_USER = 'Unban user';
+export const UNBAN_USER_SUCCESS = 'Unban user success';
+export const UNBAN_USER_FAILURE = 'Unban user failure';
+
+export const DELETE_USER = 'Delete user';
+export const DELETE_USER_SUCCESS = 'Delete user success';
+export const DELETE_USER_FAILURE = 'Delete user failure';
+
 
 export class LoadUsers implements Action {
   readonly type: string = LOAD_USERS;
@@ -374,6 +388,48 @@ export class GetLicenseFailure implements Action {
   }
 }
 
+export class BanUser implements Action {
+  readonly type: string = BAN_USER;
+
+  constructor (public payload: UserBanDetails) {
+  }
+}
+
+export class BanUserSuccess implements Action {
+  readonly type: string = BAN_USER_SUCCESS;
+
+  constructor (public payload?: UserInfo) {
+  }
+}
+
+export class UnbanUser implements Action {
+  readonly type: string = UNBAN_USER;
+
+  constructor (public payload: number) {
+  }
+}
+
+export class UnbanUserSuccess implements Action {
+  readonly type: string = UNBAN_USER_SUCCESS;
+
+  constructor (public payload?: UserInfo) {
+  }
+}
+
+export class DeleteUser implements Action {
+  readonly type: string = DELETE_USER;
+
+  constructor (public payload: number) {
+  }
+}
+
+export class DeleteUserSuccess implements Action {
+  readonly type: string = DELETE_USER_SUCCESS;
+
+  constructor (public payload?: number) {
+  }
+}
+
 export type actions =
   LoadUsers |
   LoadUsersSuccess |
@@ -432,4 +488,13 @@ export type actions =
 
   GetLicense |
   GetLicenseSuccess |
-  GetLicenseFailure;
+  GetLicenseFailure |
+
+  BanUser |
+  BanUserSuccess |
+
+  UnbanUser |
+  UnbanUserSuccess |
+
+  DeleteUser |
+  DeleteUserSuccess ;